weather moscow/ progress bar

+++ app/src/main/java/com/retivov/testregistry/weather/Main.kt
@@ -8,7 +8,7 @@ data class Main(
     val pressure: Int,
     val temp: Double,
     @SerializedName("temp_max")
-    val tempMax: Double,
+    val tempMax: Int,
     @SerializedName("temp_min")
     val tempMin: Double
 )
