+++ app/src/main/java/com/retivov/testregistry/weather/Wind.kt
@@ -3,5 +3,5 @@ package com.retivov.testregistry.weather
 
 data class Wind(
     val deg: Int,
-    val speed: Double
+    val speed: Int
 )
